requirements.txt
Django==1.9.1
djangorestframework==3.3.3
